+++ LIBRARY/CARDAMOM_F/method/samplers_shared.f90
@@ -90,7 +90,7 @@ module samplers_shared
                                                          ! for MCMC search, when applied to  multivariate proposal.
                                                          ! NOTE 1: 2.38/sqrt(npars) sometimes used when applied to the Cholesky factor.
                                                          ! NOTE 2: 2.381204**2 = 5.670132
-      double precision:: N_before_mv = 10d0 ! Number of accepted proposals before attempting to build multi-variate sampler
+      integer :: N_before_mv = 10 ! Number of accepted proposals before attempting to build multi-variate sampler
    end type MCMC_OPTIONS
 
    !> Collection of info for output of the sampling run
